requirements.txt
numpy
pandas
seaborn
matplotlib
Pillow
scikit-image==0.22.0
scikit-learn
tensorflow
keras